requirements.txt
aiogram==2.10.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
python-dotenv==0.14.0
pytz==2020.1
SpeechRecognition==3.8.1
yarl==1.6.0
